Installed types for expres, mongoose, and cors. converted db.js and app.js to typescript
