+++ package.json
@@ -30,6 +30,9 @@
     "vite": "^5.3.4"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/mongoose": "^5.11.97",
     "nodemon": "^3.1.2"
   },
   "keywords": []
